Move ActiveProfile to fiptables.
