+++ calico/felix/fiptables.py
@@ -20,10 +20,86 @@ IP tables management functions.
 """
 import logging
 from subprocess import CalledProcessError
-from calico.felix.actor import Actor, actor_event
+from calico.felix.actor import Actor, actor_event, wait_and_check
+from calico.felix.frules import (CHAIN_PROFILE_PREFIX,
+                                 rules_to_chain_rewrite_lines,
+                                 profile_to_chain_name)
 from gevent import subprocess
 
 _log = logging.getLogger(__name__)
+
+
+class ActiveProfile(Actor):
+    def __init__(self, profile_id, iptables_updaters):
+        super(ActiveProfile, self).__init__()
+        self.id = profile_id
+        self._iptables_updaters = iptables_updaters
+        self._profile = None
+        """:type dict: filled in by first update"""
+        self._tag_to_ip_set_name = None
+        """:type dict[str, str]: current mapping from tag name to ipset name."""
+
+    @actor_event
+    def on_profile_update(self, profile, tag_to_ipset_name):
+        """
+        Update the programmed iptables configuration with the new
+        profile
+        """
+        assert profile["id"] == self.id
+
+        futures = self._update_chain(profile, "inbound", tag_to_ipset_name)
+        futures += self._update_chain(profile, "outbound", tag_to_ipset_name)
+        wait_and_check(futures)
+
+        self._profile = profile
+        self._tag_to_ip_set_name = tag_to_ipset_name
+
+    @actor_event
+    def remove(self):
+        """
+        Called to tell us that this profile is no longer needed.  Removes
+        our iptables configuration.
+
+        Thread safety: Caller should wait on the result of this method before
+        creating a new ActiveProfile with the same name.  Otherwise, the
+        delete calls in this method could be issued after the initialization
+        of the new profile.
+        """
+        futures = []
+        for direction in ["inbound", "outbound"]:
+            chain_name = profile_to_chain_name(direction, self.id)
+            for updater in self._iptables_updaters.values():
+                f = updater.delete_chain(chain_name, async=True)
+                futures.append(f)
+        wait_and_check(futures)
+
+        self._profile = None
+        self._tag_to_ip_set_name = None
+
+    def _update_chain(self, new_profile, direction, tag_to_ipset_name):
+        new_rules = new_profile.get(direction, [])
+        if (self._profile is None or
+            new_rules != self._profile.get(direction) or
+            tag_to_ipset_name != self._tag_to_ip_set_name):
+            _log.debug("Update to %s affects %s rules.", self.id, direction)
+            chain_name = profile_to_chain_name(direction, self.id)
+            futures = []
+            for version, ipt in self._iptables_updaters.iteritems():
+                if version == 6:
+                    _log.error("Ignoring v6")
+                    continue
+                updates = rules_to_chain_rewrite_lines(chain_name,
+                                                       new_rules,
+                                                       version,
+                                                       tag_to_ipset_name,
+                                                       on_allow="RETURN")
+                f = ipt.apply_updates("filter", [chain_name], updates,
+                                      async=True)
+                futures.append(f)
+            return futures
+        else:
+            _log.debug("Update to %s didn't affect %s rules.",
+                       self.id, direction)
 
 
 class IptablesUpdater(Actor):
